feat: model 구조 수정
model 구조를 transformers 기반으로 수정했습니다.
