--- model/model.py
@@ -1,36 +1,0 @@
-import torch
-import torch.nn as nn
-
-class GRUNet(nn.Module):
-    def __init__(self, pretrained, hidden_dim, output_dim, n_layers, bidirectional, dropout):
-        super().__init__()
-        self.pretrained = pretrained
-        embedding_dim = pretrained.config.to_dict()['hidden_size']
-        self.rnn = nn.GRU(embedding_dim,
-                          hidden_dim,
-                          num_layers = n_layers,
-                          bidirectional = bidirectional,
-                          batch_first = True,
-                          dropout = 0 if n_layers < 2 else dropout)
-        self.out = nn.Linear(hidden_dim * 2 if bidirectional else hidden_dim, output_dim)
-        self.dropout = nn.Dropout(dropout)
-        
-    def forward(self, text):
-        #text = [batch size, sent len]
-
-        with torch.no_grad():
-            embedded = self.pretrained(text)[0]
-        #embedded = [batch size, sent len, emb dim]
-        
-        _, hidden = self.rnn(embedded)
-        #hidden = [n layers * n directions, batch size, emb dim]
-        
-        if self.rnn.bidirectional:
-            hidden = self.dropout(torch.cat((hidden[-2,:,:], hidden[-1,:,:]), dim = 1))
-        else:
-            hidden = self.dropout(hidden[-1,:,:])
-        #hidden = [batch size, hid dim]
-        
-        output = self.out(hidden)
-        #output = [batch size, out dim]
-        return output
